splitter: correctly handle chunked writes
Previously a chunked write that did not neatly end the encoded line was
handled incorrectly, resulting in overly long line on the next call to
Write().

The lineSplitter has been changed keep a running count of bytes wrote,
allowing chunked writes to be split into lines of the configured length.

Fixes #32.