Update 'Date' timestamp before sending email (#51)
As discussed in
https://github.com/domodwyer/mailyak/issues/50#issuecomment-720597806
the library should automatically updated the date timestamp before
sending out emails, otherwise re-using the MailYak instance (over
longer periods of time) leads to stale timestamps.

With this patch, the timestamp is automatically updated in Send() and
MimeBuf() methods. It also adds a testcase for this scenario.