tls: support explicit TLS connections
Adds support for connecting to the SMTP host over a TLS connection, instead of
having to use plain-text and upgrading with STARTTLS (which itself is not
secure).

TLS connections are configurable with a tls.Config at initialisation time, but a
sensible default is generated if not specified.

This change makes the underlying connection & protocol exchange behaviour
pluggable, allowing for simple tests and further extensions in the
future.

Fixes #36, closes #41 and closes #48.