Streaming MIME generation (#55)
* sender/tls: stream MIME generation to SMTP host

Changes the emailSender interface and buildMime() method to support generating
and streaming the MIME email contents (including attachments) directly over the
network connection rather than buffering the MIME in memory. This should
drastically reduce the memory overhead of sending an email, particularly when
attaching files.

Previously the full email was generated and held in memory, before connecting to
the SMTP host and sending it - this meant memory usage was proportional to the
length of the email and the size of the base64-encoded attachments. After this
change and when using TLS only a small (~4kb) buffer is held in memory at any
point (used to optimise network I/O).

This improved memory usage is currently only implemented for TLS connections.

As part of this, there is now better decoupling between the email contents and
the sender implementation resulting in better/easier tests and clearer
separation of concerns.

* sender/plain: stream MIME generation to SMTP host

Adds the streaming MIME generation from the TLS sender to the plain-text SMTP
sender.

Fixes a bug where the hostname and port was used as the hostname - this caused
errors to be returned when working with local SMTP hosts that require AUTH, and
do not support STARTTLS.