fix: include content type in empty MIME parts (#62)
Changes the MIME generation code to always include the "Content-Type"
header in each of the the MIME parts, even if the part is empty.

Previously an empty plain text or HTML body could cause some SMTP
servers to reject the mail due to the missing headers. Emails that
included both HTML and plain-text conent were not affected.

Fixes #61.